requirements.txt
Pillow==8.2.0
pynput==1.7.3
six==1.16.0
